% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor_fn_utils.R
\name{create_accessors}
\alias{create_accessors}
\title{Create accessor functions}
\usage{
create_accessors(path = "data", app_nm = "RminorElevated")
}
\arguments{
\item{path}{\code{(character)} path to directory where dependencies will be saved}

\item{app_nm}{\code{(character)} name of this app}
}
\description{
Create accessor functions
}
