% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{datatable_default}
\alias{datatable_default}
\title{DT Datatable with some helpful defaults}
\usage{
datatable_default(
  data,
  rownames = FALSE,
  filename = paste0("RminorElevated_", Sys.Date()),
  export_buttons = TRUE,
  options = list(dom = "Blfrtip", buttons = list(list(extend = "excel", customizeData =
    htmlwidgets::JS("function(data) {\\n                                        for (var i = 0; i < data.body.length; i++) {\\n                                          for (var j = 0; j < data.body[i].length; j++) {\\n                                            data.body[i][j] = '‌' + data.body[i][j];\\n                                          }\\n                                        }\\n                                      }"),
    
     exportOptions = list(modifier = list(page = "all"))), list(extend = "csvHtml5",
    exportOptions = list(modifier = list(page = "all")))), responsive = TRUE, lengthMenu
    = list(c(10, 25, 50, 75, 100, -1), c("10", "25", "50", "75", "100", "All")),
    lengthChange = TRUE, pageLength = 10),
  filter = list(position = "top", clear = TRUE, plain = FALSE),
  width = "100\%",
  height = "auto",
  extensions = "Buttons",
  style = "bootstrap4",
  elementId = NULL,
  add_options,
  add_extensions,
  ...
)
}
\arguments{
\item{data}{a data object (either a matrix or a data frame)}

\item{rownames}{\code{TRUE} (show row names) or \code{FALSE} (hide row names)
or a character vector of row names; by default, the row names are displayed
in the first column of the table if exist (not \code{NULL})}

\item{filename}{Base name (no extension) for the Excel/CSV downloads. The
Full CSV button appends "_full". Defaults to a dated stem.}

\item{export_buttons}{Whether to render the native DataTables export button
bar. Set FALSE for server-side (server = TRUE) tables, which page their
data and would export only the current page — use the download module
(mod_dt_download) on those instead.}

\item{options}{a list of initialization options (see
\url{https://datatables.net/reference/option/}); the character options
wrapped in \code{\link[htmlwidgets:JS]{htmlwidgets::JS()}} will be treated as literal
JavaScript code instead of normal character strings; you can also set
options globally via \verb{[options](DT.options = list(...))}, and
global options will be merged into this \code{options} argument if set}

\item{filter}{whether/where to use column filters; \code{none}: no filters;
\code{bottom/top}: put column filters at the bottom/top of the table; range
sliders are used to filter numeric/date/time columns, select lists are used
for factor columns, and text input boxes are used for character columns; if
you want more control over the styles of filters, you can provide a named
list to this argument; see Details for more}

\item{width, height}{Width/Height in pixels (optional, defaults to automatic
sizing)}

\item{extensions}{a character vector of the names of the DataTables
extensions (\url{https://datatables.net/extensions/index})}

\item{style}{either \code{'auto'}, \code{'default'}, \code{'bootstrap'}, or
\code{'bootstrap4'}. If \code{'auto'}, and a \strong{bslib} theme is
currently active, then bootstrap styling is used in a way that "just works"
for the active theme. Otherwise,
\href{https://datatables.net/manual/styling/classes}{DataTables \code{'default'} styling} is used. If set explicitly to \code{'bootstrap'}
or \code{'bootstrap4'}, one must take care to ensure Bootstrap's HTML
dependencies (as well as Bootswatch themes, if desired) are included on the
page. Note, when set explicitly, it's the user's responsibility to ensure
that only one unique \code{style} value is used on the same page, if multiple
DT tables exist, as different styling resources may conflict with each other.}

\item{elementId}{An id for the widget (a random string by default).}

\item{add_options}{Options to add to the existing defaults}

\item{add_extensions}{Extensions to add to the existing defaults}

\item{...}{
  Arguments passed on to \code{\link[DT:datatable]{DT::datatable}}
  \describe{
    \item{\code{class}}{the CSS class(es) of the table; see
\url{https://datatables.net/manual/styling/classes}}
    \item{\code{callback}}{the body of a JavaScript callback function with the argument
\code{table} to be applied to the DataTables instance (i.e. \code{table})}
    \item{\code{colnames}}{if missing, the column names of the data; otherwise it can be
an unnamed character vector of names you want to show in the table header
instead of the default data column names; alternatively, you can provide a
\emph{named} numeric or character vector of the form \verb{'newName1' = i1, 'newName2' = i2} or \code{c('newName1' = 'oldName1', 'newName2' = 'oldName2', ...)}, where \code{newName} is the new name you want to show in
the table, and \code{i} or \code{oldName} is the index of the current
column name}
    \item{\code{container}}{a sketch of the HTML table to be filled with data cells; by
default, it is generated from \code{htmltools::tags$table()} with a table
header consisting of the column names of the data}
    \item{\code{caption}}{the table caption; a character vector or a tag object
generated from \code{htmltools::tags$caption()}}
    \item{\code{escape}}{whether to escape HTML entities in the table: \code{TRUE} means
to escape the whole table, and \code{FALSE} means not to escape it;
alternatively, you can specify numeric column indices or column names to
indicate which columns to escape, e.g. \code{1:5} (the first 5 columns),
\code{c(1, 3, 4)}, or \code{c(-1, -3)} (all columns except the first and
third), or \code{c('Species', 'Sepal.Length')}; since the row names take
the first column to display, you should add the numeric column indices by
one when using \code{rownames}}
    \item{\code{width,height}}{Width/Height in pixels (optional, defaults to automatic
sizing)}
    \item{\code{fillContainer}}{\code{TRUE} to configure the table to automatically fill
it's containing element. If the table can't fit fully into it's container
then vertical and/or horizontal scrolling of the table cells will occur.}
    \item{\code{autoHideNavigation}}{\code{TRUE} to automatically hide navigational UI
(only display the table body) when the number of total records is less
than the page size. Note, it only works on the client-side processing mode
and the \code{pageLength} option should be provided explicitly.}
    \item{\code{lazyRender}}{\code{FALSE} to render the table immediately on page load,
otherwise delay rendering until the table becomes visible.}
    \item{\code{selection}}{the row/column selection mode (single or multiple selection
or disable selection) when a table widget is rendered in a Shiny app;
alternatively, you can use a list of the form \code{list(mode = 'multiple', selected = c(1, 3, 8), target = 'row', selectable = c(-2, -3))} to
pre-select rows and control the selectable range; the element
\code{target} in the list can be \code{'column'} to enable column
selection, or \code{'row+column'} to make it possible to select both rows
and columns (click on the footer to select columns), or \code{'cell'} to
select cells. See details section for more info.}
    \item{\code{plugins}}{a character vector of the names of DataTables plug-ins
(\url{https://rstudio.github.io/DT/plugins.html}).  Note that only those
plugins supported by the \code{DT} package can be used here. You can see
the available plugins by calling \code{DT:::available_plugins()}}
    \item{\code{editable}}{\code{FALSE} to disable the table editor, or \code{TRUE} (or
\code{"cell"}) to enable editing a single cell. Alternatively, you can set
it to \code{"row"} to be able to edit a row, or \code{"column"} to edit a
column, or \code{"all"} to edit all cells on the current page of the table.
In all modes, start editing by doubleclicking on a cell. This argument can
also be a list of the form \code{list(target = TARGET, disable = list(columns = INDICES))}, where \code{TARGET} can be \code{"cell"},
\code{"row"}, \code{"column"}, or \code{"all"}, and \code{INDICES} is an
integer vector of column indices. Use the list form if you want to disable
editing certain columns. You can also restrict the editing to accept only
numbers by setting this argument to a list of the form \code{list(target = TARGET, numeric = INDICES)} where \code{INDICES} can be the vector of the
indices of the columns for which you want to restrict the editing to
numbers or \code{"all"} to restrict the editing to numbers for all columns.
If you don't set \code{numeric}, then the editing is restricted to numbers
for all numeric columns; set \code{numeric = "none"} to disable this
behavior. It is also possible to edit the cells in text areas, which are
useful for large contents. For that, set the \code{editable} argument to a
list of the form \code{list(target = TARGET, area = INDICES)} where
\code{INDICES} can be the vector of the indices of the columns for which
you want the text areas, or \code{"all"} if you want the text areas for
all columns. Of course, you can request the numeric editing for some
columns and the text areas for some other columns by setting
\code{editable} to a list of the form \code{list(target = TARGET, numeric = INDICES1, area = INDICES2)}. Finally, you can edit date cells with a
calendar with \code{list(target = TARGET, date = INDICES)}; the target
columns must have the \code{Date} type. If you don't set \code{date} in
the \code{editable} list, the editing with the calendar is automatically
set for all \code{Date} columns.}
  }}
}
\value{
\code{(shiny.tag)}
}
\description{
DT Datatable with some helpful defaults
}
