% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dq.R
\name{dq_select_cols}
\alias{dq_select_cols}
\title{Select default display columns for Data Quality Tables}
\usage{
dq_select_cols(
  x,
  ...,
  default = list("UniqueID", "EnrollmentID", `Entry Date` = "EntryDate", "Type", "Issue")
)
}
\arguments{
\item{x}{\code{(data.frame)} The data frame from which columns will be selected.}

\item{...}{\code{(columns to select)} Columns to select, can be unquoted or quoted.}

\item{default}{\code{(list or logical)} Columns to select as defaults. If `TRUE`, defaults are used. If `FALSE`, no default columns are selected. Can also be a list of column names or expressions.}
}
\value{
\code{(data.frame)} with selected columns.
}
\description{
Select default display columns for Data Quality Tables
}
\examples{
# Selecting columns with default columns
dq_select_cols(data.frame(UniqueID = 1:3, EnrollmentID = 4:6, Issue = letters[1:3], EntryDate = 1:3, Type = 1:3))

# Selecting columns with custom defaults
dq_select_cols(data.frame(UniqueID = 1:3, EnrollmentID = 4:6, Issue = letters[1:3], EntryDate = 1:3, blah = 1:3), default = list("UniqueID", "EntryDate"))
}
