% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui.R
\name{ui_icons}
\alias{ui_icons}
\title{Iterative generation of icons}
\usage{
ui_icons(name, class = NULL, lib = "font-awesome", ...)
}
\arguments{
\item{name}{\code{(character)} The name(s) of the icon(s) to generate.}

\item{class}{\code{(character, optional)} Additional CSS class(es) to apply to the icon(s).}

\item{lib}{\code{(character)} The library from which to source the icon(s). Default is "font-awesome".}

\item{...}{Additional arguments passed to \code{\link[shiny]{icon}}.}
}
\value{
A \code{tibble} with a column of generated icons as \code{shiny::icon} objects.
}
\description{
Generates a list of icons based on the provided parameters using `shiny::icon`.
}
\examples{
# Generate icons with default parameters
ui_icons(name = c("home", "user", "cog"))
}
