% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_qpr.R
\name{mod_qpr_ui}
\alias{mod_qpr_ui}
\title{QPR UI Function}
\usage{
mod_qpr_ui(
  id,
  choices = NULL,
  date_choices = NULL,
  ns = rlang::caller_env()$ns
)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{choices}{\code{(named list)} of arguments to \link[shinyWidgets]{pickerInput}.
 `FALSE` to omit the choice drop-down selector. - choices must be provided 
 for the picker to show options. Defaults will be used for omitted arguments 
 unless arguments are explicitly set to `NULL`. Defaults are as follows:
`list(
inputId = ns("region"),
label = "Select Region(s)",
choices = choices$choices, # 
options = shinyWidgets::pickerOptions(liveSearch = TRUE),
selected = NULL,
width = "70%"
)`
Default value for choices is found in *global.R* for each tabItem namespace.}

\item{date_choices}{\code{(logical)} whether to show a date range picker}
}
\description{
A shiny Module to generate the QPR UI.
}
\seealso{
Other QPR: 
\code{\link{mod_qpr_server}()}
}
\concept{QPR}
